pyproject.toml
[project]
name = "utils"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "openpyxl>=3.1.5",
    "pandas>=2.0.0,<3.0.0",
    "streamlit>=1.30.0",
]

[dependency-groups]
dev = [
    "python-dotenv>=1.0.0",
    "pytest>=9.0.2",
]
